Develop: divisió dels fitxers per diferenciar funcionalitats, lògica i interface.
